Make FileEntry fields private, provide accessor methods
This is in preparation for Cow-ification. We want the types of those
fields to be stable no matter what the internal representation.
