+++ README.md
@@ -117,13 +117,13 @@ let pkg = rpm::Package::open("tests/assets/RPMS/v6/rpm-basic-2.3.4-5.el9.noarch.
 
 // List file metadata without reading the payload
 for entry in pkg.metadata.get_file_entries()? {
-    println!("{} ({} bytes, {:o})", entry.path.display(), entry.size, entry.mode.permissions());
+    println!("{} ({} bytes, {:o})", entry.path().display(), entry.size(), entry.permissions());
 }
 
 // Iterate over file contents (decompresses the payload)
 for entry in pkg.files()? {
     let file = entry?;
-    println!("{}: {} bytes", file.metadata.path.display(), file.content.len());
+    println!("{}: {} bytes", file.metadata.path().display(), file.content.len());
 }
 ```
