+++ src/lib.rs
@@ -108,13 +108,13 @@
 //!
 //! // List file metadata without reading the payload
 //! for entry in pkg.metadata.get_file_entries()? {
-//!     println!("{} ({} bytes, {:o})", entry.path.display(), entry.size, entry.mode.permissions());
+//!     println!("{} ({} bytes, {:o})", entry.path().display(), entry.size(), entry.permissions());
 //! }
 //!
 //! // Iterate over file contents (decompresses the payload)
 //! for entry in pkg.files()? {
 //!     let file = entry?;
-//!     println!("{}: {} bytes", file.metadata.path.display(), file.content.len());
+//!     println!("{}: {} bytes", file.metadata.path().display(), file.content.len());
 //! }
 //! # Ok(())
 //! # }
